requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
polygon-api-client>=1.14.0
python-dotenv>=1.0.0
jupyter>=1.0.0
seaborn>=0.12.0
tqdm>=4.65.0
